feat: 권한 보안 강화

+++ src/firebase/firebaseConfig.js
@@ -28,6 +28,36 @@ const app = initializeApp(firebaseConfig);
 export const db = getFirestore(app);
 export const storage = getStorage(app);
 
+export async function isOwner({ id, user }) {
+  const querySnapshot = await getDocs(
+    query(collection(db, "folder"), where("id", "==", id)),
+  );
+  let checker = true;
+
+  for (const doc of querySnapshot.docs) {
+    if (doc.data().user !== user) {
+      checker = false;
+      break;
+    }
+  }
+  return checker;
+}
+
+export async function isTextOwner({ id, user }) {
+  const querySnapshot = await getDocs(
+    query(collection(db, "text"), where("id", "==", id)),
+  );
+  let checker = true;
+
+  for (const doc of querySnapshot.docs) {
+    if (doc.data().user !== user) {
+      checker = false;
+      break;
+    }
+  }
+  return checker;
+}
+
 export async function addText({
   title,
   path,
@@ -37,27 +67,31 @@ export async function addText({
   liked,
   parentId,
 }) {
-  const newReplay = doc(collection(db, "text"));
-  await setDoc(newReplay, {
-    id: newReplay.id,
-    title,
-    realTitle,
-    path,
-    liked,
-    parentId,
-    order,
-    user,
-  });
-  return {
-    id: newReplay.id,
-    title,
-    path,
-    order,
-    realTitle,
-    liked,
-    parentId,
-    user,
-  };
+  if ((await isOwner({ id: parentId, user })) === true) {
+    const newReplay = doc(collection(db, "text"));
+    await setDoc(newReplay, {
+      id: newReplay.id,
+      title,
+      realTitle,
+      path,
+      liked,
+      parentId,
+      order,
+      user,
+    });
+    return {
+      id: newReplay.id,
+      title,
+      path,
+      order,
+      realTitle,
+      liked,
+      parentId,
+      user,
+    };
+  } else {
+    return {};
+  }
 }
 
 export async function addFolder({
@@ -68,35 +102,46 @@ export async function addFolder({
   liked,
   parentId,
 }) {
-  const newReplay = doc(collection(db, "folder"));
-  await setDoc(newReplay, {
-    id: newReplay.id,
-    title,
-    realTitle,
-    liked,
-    parentId,
-    order,
-    user,
-  });
-  return {
-    id: newReplay.id,
-    title,
-    order,
-    realTitle,
-    liked,
-    parentId,
-    user,
-  };
+  if ((await isOwner({ id: parentId, user })) === true) {
+    const newReplay = doc(collection(db, "folder"));
+    await setDoc(newReplay, {
+      id: newReplay.id,
+      title,
+      realTitle,
+      liked,
+      parentId,
+      order,
+      user,
+    });
+    return {
+      id: newReplay.id,
+      title,
+      order,
+      realTitle,
+      liked,
+      parentId,
+      user,
+    };
+  } else {
+    return {};
+  }
 }
 
 export async function getTexts(id, textId) {
-  const querySnapshot = await getDocs(
-    query(
-      collection(db, "text"),
-      where("user", "==", id),
-      where("parentId", "==", textId),
-    ),
-  );
+  let querySnapshot;
+  if (textId == "0") {
+    querySnapshot = await getDocs(
+      query(
+        collection(db, "text"),
+        where("user", "==", id),
+        where("parentId", "==", textId),
+      ),
+    );
+  } else {
+    querySnapshot = await getDocs(
+      query(collection(db, "text"), where("parentId", "==", textId)),
+    );
+  }
   if (querySnapshot.empty) {
     return [];
   }
@@ -118,13 +163,22 @@ export async function getTexts(id, textId) {
 }
 
 export async function getFolders(id, folderId) {
-  const querySnapshot = await getDocs(
-    query(
-      collection(db, "folder"),
-      where("user", "==", id),
-      where("parentId", "==", folderId || "0"),
-    ),
-  );
+  let querySnapshot;
+  console.log(id);
+  console.log(folderId);
+  if (folderId == "0") {
+    querySnapshot = await getDocs(
+      query(
+        collection(db, "folder"),
+        where("user", "==", id),
+        where("parentId", "==", folderId || "0"),
+      ),
+    );
+  } else {
+    querySnapshot = await getDocs(
+      query(collection(db, "folder"), where("parentId", "==", folderId || "0")),
+    );
+  }
   if (querySnapshot.empty) {
     return [];
   }
@@ -177,17 +231,34 @@ export async function getSpecificFolder(id) {
   const fetchedFolder = [];
   querySnapshot.forEach((doc) => {
     fetchedFolder.push({
-      id: doc.id,
-      title: doc.data()["title"],
-      liked: doc.data()["liked"],
       parentId: doc.data()["parentId"],
-      realTitle: doc.data()["realTitle"],
-      order: doc.data()["order"],
       user: doc.data()["user"],
     });
   });
   return fetchedFolder;
 }
+
+// export async function getSpecificFolder(id) {
+//   const querySnapshot = await getDocs(
+//     query(collection(db, "folder"), where("id", "==", id)),
+//   );
+//   if (querySnapshot.empty) {
+//     return [];
+//   }
+//   const fetchedFolder = [];
+//   querySnapshot.forEach((doc) => {
+//     fetchedFolder.push({
+//       id: doc.id,
+//       title: doc.data()["title"],
+//       liked: doc.data()["liked"],
+//       parentId: doc.data()["parentId"],
+//       realTitle: doc.data()["realTitle"],
+//       order: doc.data()["order"],
+//       user: doc.data()["user"],
+//     });
+//   });
+//   return fetchedFolder;
+// }
 
 export async function getChildren(id, user) {
   const querySnapshot = await getDocs(
@@ -247,16 +318,30 @@ export async function editPath({ id, type, newPath }) {
   return fetched;
 }
 
-export async function deleteSpecificText(id, title) {
-  const fileRef = ref(storage, `texts/${title}.txt`);
-  await deleteDoc(doc(db, "text", id));
-  await deleteObject(fileRef);
-  return { status: "성공" };
+export async function deleteSpecificText({ id, title, email }) {
+  if ((await isTextOwner({ id, user: email })) === true) {
+    const fileRef = ref(storage, `texts/${title}.txt`);
+    if (id !== "nope") {
+      await deleteDoc(doc(db, "text", id));
+    }
+    await deleteObject(fileRef);
+    return { status: "성공" };
+  } else {
+    if (id === "nope") {
+      const fileRef = ref(storage, `texts/${title}.txt`);
+      await deleteObject(fileRef);
+    }
+    return { status: "실패" };
+  }
 }
 
-export async function deleteSpecificFolder(id) {
-  await deleteDoc(doc(db, "folder", id));
-  return { status: "성공" };
+export async function deleteSpecificFolder({ id, email }) {
+  if ((await isOwner({ id, user: email })) === true) {
+    await deleteDoc(doc(db, "folder", id));
+    return { status: "성공" };
+  } else {
+    return { status: "실패" };
+  }
 }
 
 export async function addOrNotMember({ email, password, image }) {
